+++ simulator/src/test/java/dev/nkap/simulator/TransferApiTest.java
@@ -33,7 +33,7 @@ class TransferApiTest {
     MockMvc mvc;
 
     @Autowired
-    CollectionRequestStore store;
+    ReferenceStore store;
 
     @BeforeEach
     void reset() {